Refactored Demos
