+++ ScarletEngine/Sources/AchillesPhysics/Public/AchillesCore.h
@@ -5,3 +5,16 @@
 #include "SpringSystem.h"
 #include "RigidBodySystem.h"
 #include "ColliderSystem.h"
+
+namespace ScarletEngine::Achilles
+{
+	void RegisterSystems()
+	{
+		SystemScheduler::Get().RegisterSystem<SpringSystem>();
+		SystemScheduler::Get().RegisterSystem<RigidBodySystem>();
+		SystemScheduler::Get().RegisterSystem<AABBvsAABBColliderSystem>();
+		SystemScheduler::Get().RegisterSystem<SphereVsSphereColliderSystem>();
+		SystemScheduler::Get().RegisterSystem<PlaneVsSphereColliderSystem>();
+	}
+}
+
